feat #73: add mapper methods
